fix(trusty-installer): split verify_launchd_state out of verify_tail.rs (500-SLOC cap)
CI's "500-line file-size cap" check failed on PR #3836's rebased head:
verify_tail.rs grew to 576 SLOC (over the 500 production cap) once the
#3833 poll/aggregate-budget machinery and the #3836 down-state diagnosis
landed. Extracted the three-state launchd diagnosis (DownState,
classify_down_state, is_label_loaded, and the launchctl-list parsing it's
built on) into a new sibling module, verify_launchd_state.rs — a cohesive
concern already shared by both verify_tail::verify_one (#3833) and
service_bootstrap::bootstrap_one (#3836's defensive fallback), so it's a
natural module boundary, not an arbitrary split.

verify_tail.rs: 576 -> 425 SLOC. verify_launchd_state.rs: 157 SLOC (new).
service_bootstrap.rs: 393 SLOC (its one call site updated to the new
module path). `scripts/check_line_cap.sh` now passes locally (0 violations).
Test count unchanged (454) — every test moved with its code, none dropped.
