+++ crates/trusty-installer/src/commands/verify_launchd_state.rs
@@ -1,0 +1,335 @@
+//! Three-state launchd diagnosis for a `down` service member (#3833/#3836).
+//!
+//! Why: split out of `verify_tail.rs` (issue #3836 code-critic fix round —
+//! `verify_tail.rs` crossed the 500-SLOC production file cap once this
+//! diagnosis machinery + the #3836 defensive-fallback support landed).
+//! Distinguishing WHY a launchd-managed daemon is `down` — never loaded at
+//! all, loaded then crashed, or loaded and just still starting — is its own
+//! cohesive concern, independent of `verify_tail`'s polling/report-building
+//! logic, and is reused by BOTH `verify_tail::verify_one` (the #3833 poll
+//! wait's final diagnosis) and `service_bootstrap::bootstrap_one` (the #3836
+//! defensive fallback's "is it actually loaded?" check) — a natural module
+//! boundary.
+//!
+//! What: [`DownState`] is the three-variant diagnosis; [`classify_down_state`]
+//! derives it for a binary via `launchctl list <label>`; [`is_label_loaded`]
+//! answers the narrower "is it loaded at all?" question the #3836 fallback
+//! needs, off the SAME `launchctl list` primitive so the two call sites can
+//! never disagree about what "loaded" means.
+//!
+//! Test: the pure decision pieces (`classify_down_state_from_entry`,
+//! `parse_launchd_list_text`, `DownState::phrase`) are unit-tested; the
+//! `launchctl` subprocess call (`launchd_list_raw`) is side-effecting and
+//! validated manually.
+
+use serde::Serialize;
+
+/// Why a LAUNCHD daemon member is still reporting `down` after the #3833 poll
+/// wait — replaces a uniform, underspecified `down` with a diagnosis an
+/// operator can act on directly.
+///
+/// Why: "down" alone doesn't tell you whether launchd never loaded the job at
+/// all, loaded it and it crashed, or it is simply still coming up — three
+/// situations with three different next actions (re-run the bootstrap step,
+/// read the crash log, or just wait longer).
+/// What: [`DownState::NotLoaded`] — `launchctl list <label>` found no such
+/// service (never bootstrapped, or booted out). [`DownState::Crashed`] —
+/// loaded, no PID currently running, and the last recorded exit status was
+/// nonzero. [`DownState::StillStarting`] — loaded, and either a PID is
+/// currently running (health just hasn't caught up yet) or the last exit
+/// status was a clean `0` (about to be (re)launched, e.g. between
+/// `ThrottleInterval` respawns).
+/// Test: `tests::classify_down_state_from_entry_*`.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize)]
+#[serde(rename_all = "snake_case", tag = "state")]
+pub enum DownState {
+    /// `launchctl` has no record of this label at all.
+    NotLoaded,
+    /// Loaded, but not currently running, with a nonzero last exit status.
+    Crashed {
+        /// The `LastExitStatus` launchd recorded.
+        exit_code: i32,
+    },
+    /// Loaded and either currently running (health hasn't caught up) or
+    /// cleanly exited and awaiting its next (re)launch.
+    StillStarting,
+}
+
+impl DownState {
+    /// A short, human-readable phrase for `verify_tail`'s per-row annotation.
+    ///
+    /// Why: centralises the wording so the JSON `state` tag and the human
+    /// narration can never drift.
+    /// What: maps each variant to a lowercase phrase with no leading article.
+    /// Test: `tests::down_state_phrase_mapping`.
+    pub(super) fn phrase(&self) -> String {
+        match self {
+            DownState::NotLoaded => "not loaded".to_owned(),
+            DownState::Crashed { exit_code } => format!("crashed, exit {exit_code}"),
+            DownState::StillStarting => "still starting".to_owned(),
+        }
+    }
+}
+
+/// One `launchctl list <label>` observation, pre-parsed (#3833).
+///
+/// Why: separating the parse from the subprocess spawn lets
+/// [`classify_down_state_from_entry`] be exercised with fixed sample text —
+/// no live `launchctl` needed.
+/// What: `has_pid` — whether the dump has a `"PID" = N;` line (a process is
+/// currently running); `exit_status` — the `"LastExitStatus"` launchd last
+/// recorded (`0` when absent, matching launchd's own default).
+/// Test: `tests::parse_launchd_list_text_*`.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+struct LaunchdListEntry {
+    has_pid: bool,
+    exit_status: i32,
+}
+
+/// Parse `launchctl list <label>`'s property-list-style stdout dump.
+///
+/// Why: isolates the (pure) text parsing from the (side-effecting) subprocess
+/// spawn so the format assumptions are unit-testable.
+/// What: scans for a `"PID" = ...;` line (sets `has_pid`) and a
+/// `"LastExitStatus" = N;` line (sets `exit_status`, defaulting to `0` when
+/// absent — launchd omits the key before the job has ever exited). Never
+/// panics on malformed input.
+/// Test: `tests::parse_launchd_list_text_running`,
+/// `tests::parse_launchd_list_text_crashed`,
+/// `tests::parse_launchd_list_text_clean_exit`.
+fn parse_launchd_list_text(text: &str) -> LaunchdListEntry {
+    let has_pid = text.lines().any(|l| l.trim_start().starts_with("\"PID\""));
+    let exit_status = text
+        .lines()
+        .find_map(|l| l.trim().strip_prefix("\"LastExitStatus\" ="))
+        .and_then(|rest| rest.trim().trim_end_matches(';').trim().parse::<i32>().ok())
+        .unwrap_or(0);
+    LaunchdListEntry {
+        has_pid,
+        exit_status,
+    }
+}
+
+/// Classify a parsed `launchctl list` observation into a [`DownState`].
+///
+/// Why: the pure decision half of [`classify_down_state`] — kept separate so
+/// every branch is unit-tested without a live `launchctl`.
+/// What: `None` (label not found) → [`DownState::NotLoaded`]; a running PID →
+/// [`DownState::StillStarting`] (loaded, health just hasn't caught up); a
+/// nonzero last exit status with no running PID → [`DownState::Crashed`];
+/// anything else (loaded, no PID, clean last exit) → [`DownState::StillStarting`]
+/// (about to be (re)launched).
+/// Test: `tests::classify_down_state_from_entry_not_loaded`,
+/// `tests::classify_down_state_from_entry_running`,
+/// `tests::classify_down_state_from_entry_crashed`,
+/// `tests::classify_down_state_from_entry_clean_exit_no_pid`.
+fn classify_down_state_from_entry(entry: Option<LaunchdListEntry>) -> DownState {
+    match entry {
+        None => DownState::NotLoaded,
+        Some(e) if e.has_pid => DownState::StillStarting,
+        Some(e) if e.exit_status != 0 => DownState::Crashed {
+            exit_code: e.exit_status,
+        },
+        Some(_) => DownState::StillStarting,
+    }
+}
+
+/// Run `launchctl list <label>` and return its raw stdout, or `None` when the
+/// label is not found (macOS only; #3833).
+///
+/// Why: isolated as its own thin side-effecting function so
+/// [`classify_down_state`] composes it with the pure
+/// [`parse_launchd_list_text`] / [`classify_down_state_from_entry`] pair.
+/// What: `Some(stdout)` on a successful `launchctl list <label>`; `None` when
+/// the command fails to spawn or exits non-zero (launchd's "no such service"
+/// signature).
+/// Test: side-effecting; not invoked in the test suite.
+#[cfg(target_os = "macos")]
+fn launchd_list_raw(label: &str) -> Option<String> {
+    let out = std::process::Command::new("launchctl")
+        .args(["list", label])
+        .output()
+        .ok()?;
+    if !out.status.success() {
+        return None;
+    }
+    Some(String::from_utf8_lossy(&out.stdout).into_owned())
+}
+
+#[cfg(not(target_os = "macos"))]
+fn launchd_list_raw(_label: &str) -> Option<String> {
+    None
+}
+
+/// Whether launchd currently has `label` loaded at all (#3836).
+///
+/// Why: `service_bootstrap::bootstrap_one`'s #3836 defensive fallback needs
+/// to know whether a component binary's own `service install` actually
+/// loaded the agent, not just that the subprocess exited 0 (#3832's root
+/// cause — trusty-memory's `service install` used to write the plist without
+/// loading it). Reuses the exact same `launchctl list <label>` primitive
+/// [`classify_down_state`] is built on ([`launchd_list_raw`]), so the two
+/// call sites can never disagree about what "loaded" means.
+/// What: `true` iff `launchctl list <label>` finds the label at all —
+/// regardless of whether a PID is currently running; a loaded-but-not-yet-
+/// running job is still "loaded" (launchd owns it and will start it).
+/// Test: side-effecting; not invoked in the test suite (mirrors
+/// `classify_down_state`) — the underlying `Option`-based decision is the
+/// SAME one `classify_down_state_from_entry`'s `None` branch already covers.
+pub(super) fn is_label_loaded(label: &str) -> bool {
+    launchd_list_raw(label).is_some()
+}
+
+/// Classify why a LAUNCHD member is still `down` after the poll wait (#3833).
+///
+/// Why: `verify_tail::verify_one` calls this once it has a final `down`
+/// verdict for a LAUNCHD member — composes the side-effecting `launchctl
+/// list` call with the pure parse + classify pair above.
+/// What: resolves the member's launchd label, runs `launchctl list <label>`,
+/// and classifies the result via [`classify_down_state_from_entry`].
+/// Test: side-effecting (subprocess); the decision half is
+/// `classify_down_state_from_entry`.
+pub(super) fn classify_down_state(binary: &str) -> DownState {
+    let label = super::plist_label::plist_label_for(binary);
+    let entry = launchd_list_raw(&label).map(|text| parse_launchd_list_text(&text));
+    classify_down_state_from_entry(entry)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// Why: pins the exact wording each `DownState` variant renders, since
+    /// both the human summary and (transitively, via `serde`) the `--json`
+    /// `state` tag depend on it staying stable.
+    /// What: asserts the phrase for each variant.
+    /// Test: This is the test.
+    #[test]
+    fn down_state_phrase_mapping() {
+        assert_eq!(DownState::NotLoaded.phrase(), "not loaded");
+        assert_eq!(
+            DownState::Crashed { exit_code: 78 }.phrase(),
+            "crashed, exit 78"
+        );
+        assert_eq!(DownState::StillStarting.phrase(), "still starting");
+    }
+
+    /// Why: the running-PID branch must win regardless of last exit status —
+    /// a currently-running process means the daemon IS up; `down` health
+    /// just hasn't caught up yet (e.g. still loading models).
+    /// What: an entry with `has_pid: true` classifies as `StillStarting`
+    /// even when `exit_status` is nonzero (stale from a PRIOR crash).
+    /// Test: This is the test.
+    #[test]
+    fn classify_down_state_from_entry_running() {
+        let entry = LaunchdListEntry {
+            has_pid: true,
+            exit_status: 1,
+        };
+        assert_eq!(
+            classify_down_state_from_entry(Some(entry)),
+            DownState::StillStarting
+        );
+    }
+
+    /// Why: THE #3833 core diagnosis — no running PID plus a nonzero last
+    /// exit status is a genuine crash, not a startup race.
+    /// What: asserts `Crashed` carries the exact exit code.
+    /// Test: This is the test.
+    #[test]
+    fn classify_down_state_from_entry_crashed() {
+        let entry = LaunchdListEntry {
+            has_pid: false,
+            exit_status: 2,
+        };
+        assert_eq!(
+            classify_down_state_from_entry(Some(entry)),
+            DownState::Crashed { exit_code: 2 }
+        );
+    }
+
+    /// Why: no PID + a clean (`0`) last exit is ambiguous between "never
+    /// started yet" and "cleanly stopped, about to relaunch" — both read as
+    /// "still starting" rather than a false `Crashed`.
+    /// What: asserts `StillStarting`.
+    /// Test: This is the test.
+    #[test]
+    fn classify_down_state_from_entry_clean_exit_no_pid() {
+        let entry = LaunchdListEntry {
+            has_pid: false,
+            exit_status: 0,
+        };
+        assert_eq!(
+            classify_down_state_from_entry(Some(entry)),
+            DownState::StillStarting
+        );
+    }
+
+    /// Why: THE #3832/#3833 root-cause signature — a label `launchctl list`
+    /// cannot find at all (never bootstrapped, e.g. the #3832 trusty-memory
+    /// bug) must be reported as `NotLoaded`, not lumped in with a crash.
+    /// What: asserts `None` (label not found) classifies as `NotLoaded`.
+    /// Test: This is the test.
+    #[test]
+    fn classify_down_state_from_entry_not_loaded() {
+        assert_eq!(classify_down_state_from_entry(None), DownState::NotLoaded);
+    }
+
+    /// Why: `launchctl list <label>`'s dump format is `"Key" = value;` lines
+    /// inside a `{ ... }` block; the parser must find `PID`/`LastExitStatus`
+    /// regardless of surrounding whitespace/indentation and must not mistake
+    /// unrelated keys (e.g. `"PerJobMachServices"`) for them.
+    /// What: parses a realistic "running" dump; asserts `has_pid` and the
+    /// default `exit_status` (absent key → `0`).
+    /// Test: This is the test.
+    #[test]
+    fn parse_launchd_list_text_running() {
+        let text = r#"{
+	"LimitLoadToSessionType" = "Aqua";
+	"Label" = "com.trusty.trusty-search";
+	"OnDemand" = false;
+	"LastExitStatus" = 0;
+	"PID" = 4242;
+	"Program" = "/usr/local/bin/trusty-search";
+};
+"#;
+        let entry = parse_launchd_list_text(text);
+        assert!(entry.has_pid);
+        assert_eq!(entry.exit_status, 0);
+    }
+
+    /// Why: the crash signature — no `PID` line, a nonzero `LastExitStatus`.
+    /// What: parses a realistic "crashed" dump; asserts `has_pid` is false
+    /// and `exit_status` matches.
+    /// Test: This is the test.
+    #[test]
+    fn parse_launchd_list_text_crashed() {
+        let text = r#"{
+	"LimitLoadToSessionType" = "Aqua";
+	"Label" = "com.trusty.memory";
+	"LastExitStatus" = 78;
+};
+"#;
+        let entry = parse_launchd_list_text(text);
+        assert!(!entry.has_pid);
+        assert_eq!(entry.exit_status, 78);
+    }
+
+    /// Why: a loaded-but-never-yet-run (or cleanly stopped) job omits `PID`
+    /// and reports `LastExitStatus = 0` (or omits it entirely) — must not be
+    /// misparsed as a crash.
+    /// What: parses a dump with neither key; asserts `has_pid: false`,
+    /// `exit_status: 0` (the documented default).
+    /// Test: This is the test.
+    #[test]
+    fn parse_launchd_list_text_clean_exit() {
+        let text = r#"{
+	"Label" = "com.trusty.trusty-review";
+	"OnDemand" = false;
+};
+"#;
+        let entry = parse_launchd_list_text(text);
+        assert!(!entry.has_pid);
+        assert_eq!(entry.exit_status, 0);
+    }
+}
